Refactor most of the viewer global logic into a wrapping class
